# server/src/main/resources/openapi/storyblock-v1.yaml
openapi: 3.1.0
info:
  title: StoryBlock Engine API
  version: 1.1.0
  description: Immutable, evidence-bound narrative editing contracts.
servers:
  - url: /v1
security:
  - bearerAuth: []
...
paths:
...
  /style-profiles:
    post:
      operationId: createStyleProfile
      summary: Create an immutable novel-scoped style profile
      x-required-scopes: ["style:admin"]
      parameters:
        - $ref: "#/components/parameters/IfMatchForCreation"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StyleProfileRequest"
      responses:
        "200":
          description: Idempotent replay of the style profile creation
          headers:
            ETag:
              $ref: "#/components/headers/StrongEtag"
            Location:
              $ref: "#/components/headers/Location"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StyleProfile"
        "201":
          description: Style profile created
          headers:
            ETag:
              $ref: "#/components/headers/StrongEtag"
            Location:
              $ref: "#/components/headers/Location"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StyleProfile"
        default:
          $ref: "#/components/responses/Problem"
  /style-profiles/{profileId}:
    get:
      operationId: getStyleProfile
      summary: Get an immutable style profile within its novel boundary
      x-required-scopes: ["novel:read"]
      parameters:
        - $ref: "#/components/parameters/ProfileId"
      responses:
        "200":
          description: Immutable style profile
          headers:
            ETag:
              $ref: "#/components/headers/StrongEtag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StyleProfile"
        default:
          $ref: "#/components/responses/Problem"
  /style-profiles/{profileId}/versions:
    post:
      operationId: createStyleProfileVersion
      summary: Create an immutable DRAFT profile baseline version
      x-required-scopes: ["style:admin"]
      parameters:
        - $ref: "#/components/parameters/ProfileId"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
            application/json:
              schema:
                $ref: "#/components/schemas/StyleProfileVersionContent"
      responses:
        "200":
          description: Idempotent replay of profile version creation
          headers:
            ETag:
              $ref: "#/components/headers/StrongEtag"
            Location:
              $ref: "#/components/headers/Location"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StyleProfileVersionView"
        "201":
          description: Immutable DRAFT profile version created
          headers:
            ETag:
              $ref: "#/components/headers/StrongEtag"
            Location:
              $ref: "#/components/headers/Location"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StyleProfileVersionView"
        default:
          $ref: "#/components/responses/Problem"
  /style-profiles/{profileId}/versions/{versionId}:
    get:
      operationId: getStyleProfileVersion
      summary: Get an immutable profile version and append-only lifecycle
      x-required-scopes: ["novel:read"]
      parameters:
        - $ref: "#/components/parameters/ProfileId"
        - $ref: "#/components/parameters/ProfileVersionId"
      responses:
        "200":
          description: Style profile version lifecycle view
          headers:
            ETag:
              $ref: "#/components/headers/StrongEtag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StyleProfileVersionView"
        default:
          $ref: "#/components/responses/Problem"
  /style-profiles/{profileId}/versions/{versionId}/transitions:
    post:
      operationId: transitionStyleProfileVersion
      summary: Append an explicit audited profile lifecycle transition
      x-required-scopes: ["style:admin"]
      parameters:
        - $ref: "#/components/parameters/ProfileId"
        - $ref: "#/components/parameters/ProfileVersionId"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StyleProfileTransitionRequest"
      responses:
        "200":
          description: Updated append-only lifecycle view
          headers:
            ETag:
              $ref: "#/components/headers/StrongEtag"
            Location:
              $ref: "#/components/headers/Location"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StyleProfileVersionView"
        default:
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: nv_<public-key-id>.<256-bit-secret>
  parameters:
...
    ProfileId:
      name: profileId
      in: path
      required: true
      schema:
        type: string
        pattern: '^spf_[0-9a-f-]{36}$'
    ProfileVersionId:
      name: versionId
      in: path
      required: true
      schema:
        type: string
        pattern: '^spv_[0-9a-f-]{36}$'
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: A single quoted strong canonical SHA-256 ETag.
      schema:
        type: string
        pattern: '^"sha256:[0-9a-f]{64}"$'
    IfMatchForCreation:
      name: If-Match
      in: header
      required: true
      description: Wildcard precondition for creation in a collection.
      schema:
        const: '*'
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 200
  headers:
    StrongEtag:
      description: Quoted strong ETag of the canonical revision or immutable resource view.
      required: true
      schema:
        type: string
        pattern: '^"sha256:[0-9a-f]{64}"$'
    Location:
      description: Absolute or origin-relative durable resource URI.
      required: true
      schema:
        type: string
        format: uri-reference
...
  responses:
    Problem:
      description: RFC problem response with a stable machine code
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
    Problem:
      type: object
      required: [type, title, status, code, detail, instance, request_id]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
          pattern: '^[A-Z][A-Z0-9_]{1,63}$'
        detail:
          type: string
        instance:
          type: string
        request_id:
          type: string
        current_revision_id:
          type: string
        current_etag:
          type: string
      additionalProperties: true
...
    StyleProfileRequest:
      type: object
      required: [name, scope, provenance]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        scope:
          $ref: "#/components/schemas/StyleProfileScope"
        provenance:
          type: string
          minLength: 1
          maxLength: 2000
      additionalProperties: false
    StyleProfileScope:
      type: object
      required: [novel_id, kind, subject_id]
      properties:
        novel_id:
          type: string
          pattern: '^nov_[0-9a-f-]{36}$'
        kind:
          type: string
          enum: [novel, series, character]
        subject_id:
          type: [string, "null"]
          pattern: '^[A-Za-z0-9._:@-]{1,128}$'
      additionalProperties: false
    StyleProfile:
      type: object
      required: [profile_id, name, scope, provenance, created_by, created_at]
      properties:
        profile_id:
          type: string
          pattern: '^spf_[0-9a-f-]{36}$'
        name:
          type: string
        scope:
          $ref: "#/components/schemas/StyleProfileScope"
        provenance:
          type: string
        created_by:
          type: string
          pattern: '^[A-Za-z0-9._:@-]{1,128}$'
        created_at:
          type: string
          format: date-time
      additionalProperties: false
    StyleCorpusSource:
      type: object
      required: [source_id, content_hash, kind, provenance, license, ownership]
      properties:
        source_id:
          type: string
          pattern: '^[A-Za-z0-9._:@-]{1,128}$'
        content_hash:
          type: string
          pattern: '^sha256:[0-9a-f]{64}$'
        kind:
          type: string
          enum: [owner, licensed, public_domain, generated, mixed]
        provenance:
          type: string
          minLength: 1
          maxLength: 2000
        license:
          type: string
          minLength: 1
          maxLength: 500
        ownership:
          type: string
          minLength: 1
          maxLength: 500
      additionalProperties: false
    StyleFeatureContract:
      type: object
      required: [analyzer_version, feature_schema_version, tokenizer_id, vocabulary_hash, normalizer_version, additive_smoothing_alpha, top_k]
      properties:
        analyzer_version:
          type: string
        feature_schema_version:
          type: string
        tokenizer_id:
          type: string
        vocabulary_hash:
          type: string
          pattern: '^sha256:[0-9a-f]{64}$'
        normalizer_version:
          type: string
        additive_smoothing_alpha:
          type: number
          exclusiveMinimum: 0
          maximum: 1
        top_k:
          type: integer
          minimum: 10
          maximum: 10000
      additionalProperties: false
    StyleFeatureVector:
      type: object
      required: [channel, channel_version, contract_hash, distribution, measurements, embedding]
      properties:
        channel:
          type: string
          enum: [surface, grammar, rhythm, narrative, lexical, optional_embedding]
        channel_version:
          type: string
        contract_hash:
          type: string
          pattern: '^sha256:[0-9a-f]{64}$'
        distribution:
          type: object
          additionalProperties:
            type: number
            minimum: 0
        measurements:
          type: object
          additionalProperties:
            type: number
        embedding:
          type: array
          maxItems: 4096
          items:
            type: number
      additionalProperties: false
    StyleFeatureSet:
      type: object
      required: [source_hash, contract, channels]
      properties:
        source_hash:
          type: string
          pattern: '^sha256:[0-9a-f]{64}$'
        contract:
          $ref: "#/components/schemas/StyleFeatureContract"
        channels:
          type: array
          minItems: 5
          maxItems: 6
          items:
            $ref: "#/components/schemas/StyleFeatureVector"
      additionalProperties: false
    StyleWindowConfiguration:
      type: object
      required: [operational_graphemes, operational_stride_graphemes, micro_graphemes, micro_stride_graphemes, non_overlap_enabled]
      properties:
        operational_graphemes:
          type: integer
          minimum: 2000
          maximum: 4000
        operational_stride_graphemes:
          type: integer
          minimum: 1
          maximum: 4000
        micro_graphemes:
          type: integer
          minimum: 500
          maximum: 1000
        micro_stride_graphemes:
          type: integer
          minimum: 1
          maximum: 1000
        non_overlap_enabled:
          type: boolean
      additionalProperties: false
    StyleStratum:
      type: object
      required: [kind, speaker_id]
      properties:
        kind:
          type: string
          enum: [narration, dialogue]
        speaker_id:
          type: [string, "null"]
          pattern: '^[A-Za-z0-9._:@-]{1,128}$'
      additionalProperties: false
...
    StyleChannelCalibration:
      type: object
      required: [channel, channel_version, primary_metric, reference_distances, median, mad, q95, q99]
      properties:
        channel:
          type: string
          enum: [surface, grammar, rhythm, narrative, lexical, optional_embedding]
        channel_version:
          type: string
        primary_metric:
          type: string
          enum: [jensen_shannon_distance, wasserstein_distance, robust_l1_input, cosine_distance]
        reference_distances:
          type: array
          maxItems: 1000
          items:
            type: number
            minimum: 0
        median:
          type: number
          minimum: 0
        mad:
          type: number
          minimum: 0
        q95:
          type: number
          minimum: 0
        q99:
          type: number
          minimum: 0
      additionalProperties: false
    StyleStratumCalibration:
      type: object
      required: [stratum, window_count, confidence, channels]
      properties:
        stratum:
          $ref: "#/components/schemas/StyleStratum"
        window_count:
          type: integer
          minimum: 1
          maximum: 1000
        confidence:
          type: string
          enum: [calibrated, low_confidence]
        channels:
          type: array
          minItems: 5
          maxItems: 6
          items:
            $ref: "#/components/schemas/StyleChannelCalibration"
      additionalProperties: false
    StyleCalibrationProfile:
      type: object
      required: [calibration_schema_version, target_corpus_hash, contract_hash, window_configuration_hash, strata]
      properties:
        calibration_schema_version:
          const: style-calibration-1.0.0
        target_corpus_hash:
          type: string
          pattern: '^sha256:[0-9a-f]{64}$'
        contract_hash:
          type: string
          pattern: '^sha256:[0-9a-f]{64}$'
        window_configuration_hash:
          type: string
          pattern: '^sha256:[0-9a-f]{64}$'
        strata:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/StyleStratumCalibration"
      additionalProperties: false
    StyleProfileVersionContent:
      type: object
      required: [scope, corpus_sources, feature_set, window_configuration, calibration_statistics]
      properties:
        scope:
          $ref: "#/components/schemas/StyleProfileScope"
        corpus_sources:
          type: array
          minItems: 1
          maxItems: 1000
          items:
            $ref: "#/components/schemas/StyleCorpusSource"
        feature_set:
          $ref: "#/components/schemas/StyleFeatureSet"
        window_configuration:
          $ref: "#/components/schemas/StyleWindowConfiguration"
        calibration_statistics:
          oneOf:
            - type: object
              maxProperties: 0
            - $ref: "#/components/schemas/StyleCalibrationProfile"
      additionalProperties: false
    StyleProfileVersion:
      type: object
      required: [version_id, profile_id, version, content, created_by, created_at]
      properties:
        version_id:
          type: string
          pattern: '^spv_[0-9a-f-]{36}$'
        profile_id:
          type: string
          pattern: '^spf_[0-9a-f-]{36}$'
        version:
          type: integer
          minimum: 1
        content:
          $ref: "#/components/schemas/StyleProfileVersionContent"
        created_by:
          type: string
          pattern: '^[A-Za-z0-9._:@-]{1,128}$'
        created_at:
          type: string
          format: date-time
      additionalProperties: false
    StyleLifecycleEvent:
      type: object
      required: [event_id, profile_id, version_id, sequence, from_state, to_state, reason, generated_promotion_confirmed, request_id, actor_id, actor_key_id, occurred_at]
      properties:
        event_id:
          type: string
          pattern: '^sle_[0-9a-f-]{36}$'
        profile_id:
          type: string
          pattern: '^spf_[0-9a-f-]{36}$'
        version_id:
          type: string
          pattern: '^spv_[0-9a-f-]{36}$'
        sequence:
          type: integer
          minimum: 1
        from_state:
          type: [string, "null"]
          enum: [draft, calibrating, ready, deprecated, null]
        to_state:
          type: string
          enum: [draft, calibrating, ready, deprecated]
        reason:
          type: string
          minLength: 1
          maxLength: 1000
        generated_promotion_confirmed:
          type: boolean
        request_id:
          type: string
        actor_id:
          type: string
          pattern: '^[A-Za-z0-9._:@-]{1,128}$'
        actor_key_id:
          type: [string, "null"]
          pattern: '^key_[0-9a-f-]{36}$'
        occurred_at:
          type: string
          format: date-time
      additionalProperties: false
    StyleProfileVersionView:
      type: object
      required: [profile_version, lifecycle, state, approved_by, approved_at, can_gate_rewrites]
      properties:
        profile_version:
          $ref: "#/components/schemas/StyleProfileVersion"
        lifecycle:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/StyleLifecycleEvent"
        state:
          type: string
          enum: [draft, calibrating, ready, deprecated]
        approved_by:
          type: [string, "null"]
          pattern: '^[A-Za-z0-9._:@-]{1,128}$'
        approved_at:
          type: [string, "null"]
          format: date-time
        can_gate_rewrites:
          type: boolean
      additionalProperties: false
    StyleProfileTransitionRequest:
      type: object
      required: [target_state, reason, confirm_generated_corpus_promotion]
      properties:
        target_state:
          type: string
          enum: [calibrating, ready, deprecated]
        reason:
          type: string
          minLength: 1
          maxLength: 1000
        confirm_generated_corpus_promotion:
          type: boolean
          description: Must be true when promoting generated or mixed corpus to READY.
      additionalProperties: false
...